+++ src/components/ProductCard/ProductCardWrapper.tsx
@@ -4,24 +4,36 @@ import { cnCreate } from '../../utils/cn';
 import './ProductCardWrapper.less';
 
 interface IProductCardWrapperProps {
+    /** This option add specific className on container
+     * and child components change styles: background, paddings, font-size...
+     */
+    theme?: 'default';
     /** Hint from left */
     hint?: {
         title: string;
         color: 'green' | 'orange' | 'black';
     };
-    /** Show right border */
+    /** Show border
+     * each property:'sky' or true(bool)
+     * true - get default color #D8D8D8
+     * sky - #EDEDED
+     */
     border?: Partial<{
-        top: boolean;
-        right: boolean;
-        bottom: boolean;
-        left: boolean;
+        top: boolean | 'sky';
+        right: boolean | 'sky';
+        bottom: boolean | 'sky';
+        left: boolean | 'sky';
     }>;
     children: JSX.Element[] | Element[] | JSX.Element | Element;
+    className?: string;
+    classNameInner?: string;
+    classNameContainer?: string;
 }
 
 const cn = cnCreate('product-card-wrapper');
 class ProductCardWrapper extends React.Component<IProductCardWrapperProps, {}> {
     static propTypes = {
+        theme: PropTypes.oneOf(['default']),
         hint: PropTypes.shape({
             title: PropTypes.string.isRequired,
             color: PropTypes.oneOf(['green', 'orange', 'black']).isRequired,
@@ -31,6 +43,15 @@ class ProductCardWrapper extends React.Component<IProductCardWrapperProps, {}> {
             PropTypes.element,
             PropTypes.node,
         ]),
+        border: PropTypes.shape({
+            top: PropTypes.oneOf([false, true, 'sky']),
+            right: PropTypes.oneOf([false, true, 'sky']),
+            bottom: PropTypes.oneOf([false, true, 'sky']),
+            left: PropTypes.oneOf([false, true, 'sky']),
+        }),
+        className: PropTypes.string,
+        classNameInner: PropTypes.string,
+        classNameContainer: PropTypes.string,
     };
 
     static defaultProps: Partial<IProductCardWrapperProps> = {
@@ -51,22 +72,26 @@ class ProductCardWrapper extends React.Component<IProductCardWrapperProps, {}> {
     }
 
     render() {
-        const { hint, border } = this.props;
+        const { hint, border, theme, className, classNameInner, classNameContainer } = this.props;
+        const getHintColor = () => !!hint && (hint! as {}).constructor.name === 'Object' ? hint.color : null;
+        const color = getHintColor();
 
         return (
             <div
                 className={cn('', {
-                    hint: !!hint ? hint.color : false,
-                    'bl-yes': border!.left,
-                    'br-yes': border!.right,
-                    'bt-yes': border!.top,
-                    'bb-yes': border!.bottom,
-                })}
+                    hint: !!hint,
+                    [`hint_${color}`]: !!color,
+                    bl: border!.left,
+                    br: border!.right,
+                    bt: border!.top,
+                    bb: border!.bottom,
+                    theme,
+                }, className)}
                 onClick={this.handleClick}
             >
-                <div className={cn('inner')}>
+                <div className={cn('inner', {}, classNameInner)}>
                     {hint && this.renderHintLabel()}
-                    <div className={cn('container')}>
+                    <div className={cn('container', {}, classNameContainer)}>
                         {this.props.children}
                     </div>
                 </div>